backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
pdf2image==1.16.3
pytesseract==0.3.10
qdrant-client==1.5.4
httpx==0.24.1
openai==0.28.1
python-dotenv==1.0.0
langchain==0.0.306
tiktoken==0.5.1 